@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of ContaoComponentStyleManager.
+ *
+ * (c) https://www.oveleon.de/
+ */
+
 namespace Oveleon\ContaoComponentStyleManager\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
@@ -9,7 +17,7 @@ use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 class ContaoComponentStyleManagerExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $config = $this->processConfiguration(new Configuration(), $configs);
 
