@@ -1,8 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of ContaoComponentStyleManager.
+ *
+ * (c) https://www.oveleon.de/
+ */
+
 namespace Oveleon\ContaoComponentStyleManager\EventListener\DataContainer;
 
-use Contao\CoreBundle\ServiceAnnotation\Callback;
+use Contao\CoreBundle\DependencyInjection\Attribute\AsCallback;
 use Contao\Database;
 use Contao\DataContainer;
 use Contao\StringUtil;
@@ -14,18 +22,11 @@ use Symfony\Component\Routing\RouterInterface;
 
 class StyleManagerArchiveListener
 {
-
-    private RouterInterface $router;
-
-
-    public function __construct(RouterInterface $router)
+    public function __construct(private readonly RouterInterface $router)
     {
-        $this->router = $router;
     }
 
-    /**
-     * @Callback(table="tl_style_manager_archive", target="config.onload")
-     */
+    #[AsCallback(table: 'tl_style_manager_archive', target: 'config.onload')]
     public function checkIdentifier($dc): void
     {
         $objArchive = StyleManagerArchiveModel::findById($dc->id);
@@ -37,9 +38,7 @@ class StyleManagerArchiveListener
         }
     }
 
-    /**
-     * @Callback(table="tl_style_manager_archive", target="list.label.label")
-     */
+    #[AsCallback(table: 'tl_style_manager_archive', target: 'list.label.label')]
     public function addIdentifierInfo($row, $label)
     {
         if($row['identifier'])
@@ -50,9 +49,7 @@ class StyleManagerArchiveListener
         return $label;
     }
 
-    /**
-     * @Callback(table="tl_style_manager_archive", target="list.global_operations.import.button")
-     */
+    #[AsCallback(table: 'tl_style_manager_archive', target: 'list.global_operations.import.button')]
     public function importConfigButton(?string $href, string $label, string $title, string $class, string $attributes): string
     {
         if(System::getContainer()->getParameter('contao_component_style_manager.use_bundle_config'))
@@ -72,9 +69,7 @@ class StyleManagerArchiveListener
         ]);
     }
 
-    /**
-     * @Callback(table="tl_style_manager_archive", target="list.global_operations.config.button")
-     */
+    #[AsCallback(table: 'tl_style_manager_archive', target: 'list.global_operations.config.button')]
     public function bundleConfigButton(?string $href, string $label, string $title, string $class, string $attributes): string
     {
         if(System::getContainer()->getParameter('contao_component_style_manager.use_bundle_config'))
@@ -98,9 +93,7 @@ class StyleManagerArchiveListener
         return '';
     }
 
-    /**
-     * @Callback(table="tl_style_manager_archive", target=""fields.identifier.save")
-     */
+    #[AsCallback(table: 'tl_style_manager_archive', target: 'fields.identifier.save')]
     public function generateIdentifier($varValue, DataContainer $dc): string
     {
         $aliasExists = function (string $alias) use ($dc): bool
