@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of ContaoComponentStyleManager.
  *
@@ -7,16 +10,14 @@
 
 namespace Oveleon\ContaoComponentStyleManager\EventListener;
 
+use Contao\CoreBundle\DependencyInjection\Attribute\AsHook;
 use Contao\CoreBundle\Exception\NoContentResponseException;
-use Contao\CoreBundle\ServiceAnnotation\Hook;
 use Contao\DataContainer;
 use Contao\Input;
 use Contao\System;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 
-/**
- * @Hook("executePostActions")
- */
+#[AsHook('executePostActions')]
 class ExecutePostActionsListener
 {
     /**
