@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of ContaoComponentStyleManager.
  *
@@ -44,7 +47,6 @@ use Oveleon\ContaoComponentStyleManager\StyleManager\Config;
  *
  * @author Daniele Sciannimanica <daniele@oveleon.de>
  */
-
 class StyleManagerArchiveModel extends Model
 {
 
